+++ example/lib/routes/inputs.dart
@@ -1,6 +1,6 @@
 export '../screens/inputs/button.dart';
 export '../screens/inputs/checkbox.dart';
 export '../screens/inputs/radio_button.dart';
-export '../screens/inputs/rating_bar.dart';
+export '../screens/inputs/rating_control.dart';
 export '../screens/inputs/slider.dart';
 export '../screens/inputs/toggle_switch.dart';